// services/upload-service/src/infrastructure/persistence/prisma/generated/models/MediaFile.ts
export type MediaFileWhereInput = {
  AND?: Prisma.MediaFileWhereInput | Prisma.MediaFileWhereInput[]
  OR?: Prisma.MediaFileWhereInput[]
  NOT?: Prisma.MediaFileWhereInput | Prisma.MediaFileWhereInput[]
  id?: Prisma.StringFilter<"MediaFile"> | string
  s3Key?: Prisma.StringFilter<"MediaFile"> | string
  bucket?: Prisma.StringFilter<"MediaFile"> | string
  fileSizeBytes?: Prisma.BigIntFilter<"MediaFile"> | bigint | number
  originalFilename?: Prisma.StringFilter<"MediaFile"> | string
  mimeType?: Prisma.StringFilter<"MediaFile"> | string
  uploadedBy?: Prisma.StringFilter<"MediaFile"> | string
  status?: Prisma.EnumMediaProcessingStatusFilter<"MediaFile"> | $Enums.MediaProcessingStatus
  contentId?: Prisma.StringNullableFilter<"MediaFile"> | string | null
  contentType?: Prisma.StringNullableFilter<"MediaFile"> | string | null
  streamingUrl?: Prisma.StringNullableFilter<"MediaFile"> | string | null
  trailerUrl?: Prisma.StringNullableFilter<"MediaFile"> | string | null
  downloadUrl?: Prisma.StringNullableFilter<"MediaFile"> | string | null
  processingError?: Prisma.StringNullableFilter<"MediaFile"> | string | null
  deletedAt?: Prisma.DateTimeNullableFilter<"MediaFile"> | Date | string | null
  createdAt?: Prisma.DateTimeFilter<"MediaFile"> | Date | string
  updatedAt?: Prisma.DateTimeFilter<"MediaFile"> | Date | string
  previewS3Key?: Prisma.StringNullableFilter<"MediaFile"> | string | null
  previewStatus?: Prisma.EnumMediaProcessingStatusFilter<"MediaFile"> | $Enums.MediaProcessingStatus
}